Improve AlbumProcessor specs

+++ spec/features/album/album_processor_spec.rb
@@ -4,45 +4,44 @@ describe AlbumProcessor do
   let(:base_path) { Rails.root.join("spec", "fixtures", "files", "photos") }
   let(:web_path) { base_path.join("web") }
   let(:small_path) { base_path.join("small") }
-  let(:thumb_path) { base_path.join("thumbs") }
+  let(:thumbs_path) { base_path.join("thumbs") }
   let!(:originals) { Dir.entries(base_path).select { |f| f =~ /\.jpg|png\Z/i } }
   let(:web_version_path) { web_path.join(originals.first) }
   let(:small_version_path) { small_path.join(originals.first) }
-  let(:thumb_version_path) { thumb_path.join(originals.first) }
+  let(:thumbs_version_path) { thumbs_path.join(originals.first) }
   let(:image) { double(:image).as_null_object }
+  let(:web_proc) { double(:web_proc, call: image) }
+  let(:small_proc) { double(:web_proc, call: image) }
+  let(:versions) { { web: web_proc, small: small_proc } }
   subject(:processor) { AlbumProcessor.new(base_path) }
 
   def mock_image
+    stub_const("AlbumProcessor::VERSIONS", versions)
     allow(Magick::ImageList).to receive(:new) { image }
     allow(image).to receive(:resize_to_fit) { image }
     allow(image).to receive(:resize_to_fill) { image }
   end
 
   describe "#process_all" do
-
     it "create the versions on disk" do
       processor.process_all
       expect(File.exists?(web_version_path)).to be true
       expect(File.exists?(small_version_path)).to be true
-      expect(File.exists?(thumb_version_path)).to be true
+      expect(File.exists?(thumbs_version_path)).to be true
     end
 
     it "skips processed photos" do
       processor.create_versions(:web)
-      processor.create_versions(:thumbs)
       mock_image
       processor.process_all
-      expect(image).to have_received(:resize_to_fill).with(320, 320).once
-      expect(image).not_to have_received(:resize_to_fill).with(75, 75)
-      expect(image).not_to have_received(:resize_to_fit)
+      expect(small_proc).to have_received(:call).with(image).once
     end
 
     it "resizes images for versions" do
       mock_image
       processor.process_all
-      expect(image).to have_received(:resize_to_fit).with(1024, 1024).once
-      expect(image).to have_received(:resize_to_fill).with(320, 320).once
-      expect(image).to have_received(:resize_to_fill).with(75, 75).once
+      expect(web_proc).to have_received(:call).with(image).once
+      expect(small_proc).to have_received(:call).with(image).once
     end
   end
 
@@ -51,20 +50,39 @@ describe AlbumProcessor do
       processor.create_versions(:web)
       mock_image
       processor.create_versions(:web)
-      expect(image).not_to have_received(:resize_to_fit)
+      expect(web_proc).not_to have_received(:call)
     end
 
     it "does not skip processed photos when force is true" do
       processor.create_versions(:web)
       mock_image
       processor.create_versions(:web, force: true)
-      expect(image).to have_received(:resize_to_fit)
+      expect(web_proc).to have_received(:call).with(image)
+    end
+  end
+
+  describe "::VERSIONS" do
+    subject(:versions) { AlbumProcessor::VERSIONS }
+
+    it "web resizes to fit" do
+      versions[:web].call(image)
+      expect(image).to have_received(:resize_to_fit).with(1024, 1024)
+    end
+
+    it "small resizes to fill" do
+      versions[:small].call(image)
+      expect(image).to have_received(:resize_to_fill).with(320, 320)
+    end
+
+    it "thumbs resizes to fill" do
+      versions[:thumbs].call(image)
+      expect(image).to have_received(:resize_to_fill).with(75, 75)
     end
   end
 
   after do
-    FileUtils.remove_dir web_path
-    FileUtils.remove_dir small_path
-    FileUtils.remove_dir thumb_path
+    FileUtils.remove_dir web_path if File.exists? web_path
+    FileUtils.remove_dir small_path if File.exists? small_path
+    FileUtils.remove_dir thumbs_path if File.exists? thumbs_path
   end
 end
